requirements.txt
langchain
langchain-openai
langchain-community
tiktoken
python-dotenv
faiss-cpu
streamlit
youtube-transcript-api
yt_dlp
openpyxl
python-dotenv